Reserve only caml_minor_heap_max_wsz * Max_domains for the minor heap area
When a domain uses Gc.set to increase its own minor heap size to be larger
than the current reservation, we reserve a new memory area for
all minor heaps. This is done inside a STW section:

1. every domain frees its own heap,
2. inside a global barrier, the final domain unreserves the old minor heap
area, then reserves the new minor heap area. This domain also adjusts
the minor heap area pointers of all domains while the other domains
wait for the end of the barrier,
3. every domain allocates its own heap with size `Caml_state->minor_heap_wsz`.

Co-authored-by: Gabriel Scherer <gabriel.scherer@gmail.com>