Gc.set now changes the minor heap size of all domains
Every call to `Gc.set` that modifies the minor heap size causes
a stop-the-world section, in order to adjust the minor heaps of all
domains.

On `domain_terminate()`, we free the minor heap to maintain the
invariant "only active domains have a minor heap allocated".

Resulting simplifications: we no longer keep track of individual
domains' minor heap size - instead, we have a global variable
`caml_minor_heap_wsz` for the size of the minor heap, and
every domain allocates a minor heap of this size.

This patch resolves the FIXME from #11082 relating to the duplicate
reallocation of the leader domain's minor heap.